requirements.txt
accelerate==0.33.0
aiohttp==3.9.5
anthropic==0.47.2
beautifulsoup4==4.13.3
cairosvg==2.7.1
cohere==5.13.12
colorama==0.4.6
dashinfer==1.3.0
datasets==3.0.0
deepspeed==0.16.4
dspy==2.5.43
exllamav2==0.2.8
fastapi==0.115.8
flash_attn==2.7.4.post1
google_cloud_aiplatform==1.71.1
gradio==5.18.0
hdbscan==0.8.40
httpx==0.28.1
huggingface_hub==0.24.6
hydra-core==1.3.2
intel_extension_for_pytorch==2.6.0
jsonlines==4.0.0
kaleido==0.2.1.post1
lightllm==0.0.1
litellm==1.61.17
llama==0.1.1
markdownify==1.0.0
matplotlib==3.10.0
mistralai==1.5.0
mlx==0.23.1
mlx_lm==0.21.4
modelscope==1.23.1
nltk==3.9.1
numpy==2.2.3
openai==1.64.0
opencc==1.1.9
orjson==3.10.15
pandas==2.2.3
peft==0.12.0
Pillow==11.1.0
plotly==6.0.0
polyglot==16.7.4
prompt_toolkit==3.0.47
protobuf==5.29.3
psutil==6.0.0
psutil==5.9.5
pycld2==0.41
pydantic==2.10.6
pydantic_settings==2.8.0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.1
PyYAML==6.0.2
ray==2.34.0
Requests==2.32.3
rich==13.9.4
rwkv==0.8.28
safetensors==0.5.3
scikit_learn==1.4.2
scipy==1.15.2
seaborn==0.13.2
sentence_transformers==3.4.1
sglang==0.4.3.post2
shortuuid==1.0.13
tiktoken==0.7.0
torch==2.4.0
torch_npu==2.4.0.post2
tqdm==4.66.5
transformers==4.47.1
uvicorn==0.34.0
vertexai==1.71.1
vllm==0.6.3
xformers==0.0.27.post2
